Description of fields with labels from contenttype
The documentation says labels of fields defined in contenttypes.yml are for the built-in editing form.
But there is no hint if it's for labeling the fields for the webpage, too. 
With this commit, you take the label from the contenttype if it's not empty else you take the key.